requirements.txt
praw
openai
python-dotenv
beautifulsoup4
requests